+++ test/backward/compat.js
@@ -1,4 +1,4 @@
-// i18next, v1.10.3
+// i18next, v1.11.0
 // Copyright (c)2015 Jan Mühlemann (jamuhl).
 // Distributed under MIT license
 // http://i18next.com
@@ -1001,6 +1001,29 @@ describe('i18next', function() {
       })
     
     });
+   
+    describe('determining language directionality', function(){
+            
+      i18n.init();
+    
+      it('returns ltr for en-US', function(){
+        i18n.setLng('en-US');
+        expect(i18n.dir()).to.equal('ltr');
+      });
+    
+      it('returns ltr for unknown language', function(){
+        i18n.setLng('unknown');
+        expect(i18n.dir()).to.equal('ltr');
+      });
+    
+      it('returns rtl for ar, ar-IR', function(){
+        i18n.setLng('ar');
+        expect(i18n.dir()).to.equal('rtl');
+        i18n.setLng('ar-IR');
+        expect(i18n.dir()).to.equal('rtl');
+      })
+    
+    });
   
     describe('preloading multiple languages', function() {
     
@@ -1063,6 +1086,10 @@ describe('i18next', function() {
     
         it('it should postprocess with multiple post processors', function() {
           expect(i18n.t('simpleTest', {postProcess: ['myProcessor', 'myProcessor2']})).to.be('ok_from_postprocessor ok');
+        });
+    
+        it('it should postprocess on missing value with multiple post processes', function() {
+          expect(i18n.t('notFound2', {postProcess: ['myProcessor', 'myProcessor2']})).to.be('ok_from_postprocessor ok');
         });
     
         describe('or setting it as default on init', function() {